Update patch set 1
Patch Set 1:

This is incremental progress towards the "Master API" that we discussed in the meeting.  This is not intended to be the final method of reporting the results.

It shows how we can aggregate the results from multiple runs into a single report by having a run ID associated with the whole series.

Patch-set: 1
Reviewer: Gerrit User 2034 <2034@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0