Add Range function for Steady State detection
Added a range_value function in utilities/math.py able to compute the range
of a series of y values : [y1, y2, ..., yn].
Implemented a test harness for this range_value function in the tests/utilities
section.
Renamed the math_slope.py and math_range.py test files to add _test.py for
Jenkins.
Cleaned up the code so it is compliant to the pep8 rules.

Renamed the previous 'math' modules (storperf/utilities/math.py
and storperf/test/utilities/math.py) as 'math_slope' to be
coherent with the new notation.

Change-Id: I02ccd2b87f0b72e7a28c416b593aae4d8ad97961
JIRA: STORPERF-57
JIRA: STORPERF-58
Signed-off-by: Tim Rault <tim.rault@cengn.ca>