Add Steady State Detection module
Added a Steady State Detection module containing a steady_state(data_series)
function that is able to return a boolean indicating wether or not steady
state is reached with the data_series being passed.
This module requires a data_treatment(data_series) and an average(data_series)
modules that have been added in this commit as well. The data treatment function
aims at formatting the data series that is passed to the high level steady_state
function to reach the requirement of each sub-module (slope, average and range).

Modified the Slope and Range functions so they return None when passed an empty
data series instead of 0 which was wrong. Modified the corresponding test cases.

Change-Id: I9c3854cb0a21cc37b0787b8afca0821eefaa203d
JIRA: STORPERF-60
JIRA: STORPERF-59
JIRA: STORPERF-61
JIRA: STORPERF-62
Signed-off-by: Tim Rault <tim.rault@cengn.ca>