Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

Unit tests are not end-to-end functional tests.  In StorPerf we do code level testing with mocks patching the real client. So, we should be able to mock the return of no stack being created from heat.

However, having said that, the rest_server.py code is not in the storperf/ subdirectory, so it cannot have unit tests written for it.  I need to fix that at some point.

Patch-set: 2
Label: Code-Review=+2