Creation of Local Docker for Daily Job
Moves the dockerfiles for the different submodules/containers into their respective
directories.
Adds a docker compose file that can build the containers from the local filesystem,
and mounts the local filesystem as a volume, making it easier for developer testing
as local changes are automatically reflected in the final container.

Change-Id: I2376034258cab2a3dc3a0d1ac3f01bbf5d1f59e7
JIRA: STORPERF-189
Signed-off-by: mbeierl <mark.beierl@dell.com>