Adds Volume Count and Dynamic Reload
Adds a volume_count parameter that allows the user to
change the number of volumes.  Test execution still
uses only one volume by name or /dev/vdb by default.

Changes the cache policy of the stack from SNAPS-OO
so that changes made to the stack are refelcted in
StorPerf.

Change-Id: I5b9160b469e80d4f7d6906ac270bb3b75411fb5a
JIRA: STORPERF-242
Signed-off-by: mbeierl <mark.beierl@dell.com>